dev-audit-clear

+++ multi-user-blog/src/app/lib/posts/deletePost.js
@@ -2,6 +2,7 @@
 
 import { redirect } from "next/navigation";
 import { db } from "../turso";
+import { getUser } from "../getUser";
 
 
 export async function deletePostServerAction(_, formData) {
@@ -11,14 +12,26 @@ export async function deletePostServerAction(_, formData) {
         return { ok: false, message: "Post public id is broken or missing." };
     }
 
+
     try {
 
-        const fetchPost = await db.execute(`SELECT id FROM posts WHERE public_id = ?`, [ppid]);
+        const currentUser = await getUser();
+        if (!currentUser?.id) {
+            return { ok: false, message: "User needs to be logged in." };
+        }
+
+
+        const fetchPost = await db.execute(`SELECT id, user_id FROM posts WHERE public_id = ?`, [ppid]);
+
         if (fetchPost.rows.length === 0) {
             return { ok: false, message: "Post not found." };
         }
 
-        const postId = fetchPost.rows[0].id;
+        if (fetchPost?.rows[0]?.user_id !== currentUser?.id) {
+            return { ok: false, message: "You are not authorized to delete this post." };
+        }
+
+        const postId = fetchPost?.rows[0]?.id;
 
         const result = await db.execute(`DELETE FROM posts WHERE id = ?`, [postId]);
 
@@ -26,7 +39,7 @@ export async function deletePostServerAction(_, formData) {
             return { ok: false, message: "failed to delete post" };
         }
 
-       //return { ok: true, message: "post deleted" };
+        //return { ok: true, message: "post deleted" };
 
     } catch (error) {
         console.error(error);
